requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
